initial upload
Exploratory Data Analysis covers statistical principles supporting the data analytics life cycle. Students in this course compute and interpret measures of central tendency, correlations, and variation. The course introduces hypothesis testing, focusing on the application for parametric tests, and addresses communication skills and tools to explain an analyst’s findings to others within an organization. Data Cleaning is a required prerequisite for this course.